backend/requirements.txt
Django==3.2.16
djangorestframework
django-colorfield
djoser